Create index.php
Created a HTML form and used post  in the php section of the code to securely get data from the form.

There are two different types of special variables in PHP associated with getting data from forms, which is get and post, post is more secure than get. Which is why I chose post.